optimize network traffic (partial regressions)
Instead of sending data every 50 ms, data is sent every time a set of
tank actions changes.
