+++ bangbang/server.py
@@ -15,123 +15,29 @@ import bbutils
 import collisions
 import constants
 import server_network
-from shapes import HeadlessMine, HeadlessShell
+from shapes import HeadlessMine, HeadlessShell, HeadlessTank
 import utils_3d
 
 # TODO: add consistent type hinting throughout the whole project
 
-
-# TODO: rename this class to HeadlessTank and move it to shapes.py
-class Tank(base_shapes.Shape, constants.Tank):
+class Tank(HeadlessTank):
     def __init__(self, angle, client_id, color, ground_hw, name, pos, server):
-        super().__init__()
+        super().__init__(angle, client_id, color, ground_hw, name, pos)
 
-        self.bangle = angle  # base angle
-        self.tangle = angle  # turret angle
-        self.client_id = client_id
-        self.color = color
-        self.ground_hw = ground_hw
-        self.name = name
-        self.pos = np.array(pos)
-
-        # snapping back: turret turns to meet base
-        # turning back: base turns to meet turret
-        # these variables are true if either t or ctrl t have been pressed
-        # I believe they can both be true simulatneously
-        self.snapping_back = False
-        self.turning_back = False
-
-        self.hits_left: int = self.HITS_TO_DIE  # how many more hits before dead?
         self.mine_reloading = 0  # timestamp at which tank can lay a mine again
         self.shell_reloading = 0  # timestamp at which tank can fire a shell again
-        self.speed = 0.0  # m / s, I hope
 
-        self.actions: set[constants.Action] = set()
-
+        # whether a tank state change has occurred that needs to be sent to the clients
+        self.__needs_update = False
         self.server = server
 
-    def recv_hit(self, damage: int) -> None:
-        """Decrement the tank health by damage."""
-        self.hits_left -= damage
-        if self.hits_left <= 0:
-            self.die()
-
-    def snap_logic(self, target_angle, approaching_angle, incr):
-        """
-        Perform calculations for the snapping and turning back animations.
-
-        Returns the increment to be added to the approaching angle (which may be 0)
-        and a boolean indicating whether the animation is still going.
-
-        incr should be delta * (self.SNAP_SPEED or self.BROTATE).
-        """
-        diff = approaching_angle - target_angle
-        # the sign of the increment
-        direction = (diff < 0) - (diff > 0)
-        diff = abs(diff)
-        # always choose the shortest path
-        if diff > 180.0:
-            diff = 360 - diff
-            direction *= -1
-
-        # once within a certain threshold of the correct angle, stop snapping back
-        if diff <= incr:
-            return diff * direction, False
-        return incr * direction, True
-
-    def update(self):
-        delta = self.delta_time()
-
-        ip_bangle = 0
-        ip_tangle = 0
-
-        # TODO: ask some knowledgeable folks whether a match-case pattern would be more
-        # appropriate here
-        if constants.Action.ACCEL in self.actions:
-            self.speed = min(self.speed + self.ACC * delta, self.MAX_SPEED)
-
-        if constants.Action.ALL_LEFT in self.actions:
-            self.turning_back = False
-            self.snapping_back = False
-            ip_bangle = self.BROTATE
-            ip_tangle = self.BROTATE
-
-        if constants.Action.ALL_RIGHT in self.actions:
-            self.turning_back = False
-            self.snapping_back = False
-            ip_bangle = -self.BROTATE
-            ip_tangle = -self.BROTATE
-
-        if constants.Action.BASE_LEFT in self.actions:
-            # cancel turning back upon manual turn
-            self.turning_back = False
-            ip_bangle = self.BROTATE
-
-        if constants.Action.BASE_RIGHT in self.actions:
-            # cancel turning back upon manual turn
-            self.turning_back = False
-            ip_bangle = -self.BROTATE
-
-        if constants.Action.DEACCEL in self.actions:
-            self.speed = max(self.speed - self.ACC * delta, self.MIN_SPEED)
+    def update(self) -> bool:
+        super().update()
 
         if constants.Action.MINE in self.actions:
             if self.clock >= self.mine_reloading + constants.Mine.RELOAD_TIME:
                 self.mine_reloading = self.clock
                 self.server.make_mine(self.client_id, self.pos)
-
-        if constants.Action.TURN_BACK in self.actions:
-            self.turning_back = True
-
-        if constants.Action.TURRET_LEFT in self.actions:
-            # cancel snapping back upon manual turn
-            self.snapping_back = False
-            ip_tangle = self.TROTATE
-
-        if constants.Action.TURRET_RIGHT in self.actions:
-            # cancel snapping back upon manual turn
-            self.snapping_back = False
-            ip_tangle = -self.TROTATE
 
         if constants.Action.SHELL in self.actions:
             if self.clock >= self.shell_reloading + constants.Shell.RELOAD_TIME:
@@ -143,75 +49,13 @@ class Tank(base_shapes.Shape, constants.Tank):
                     self.pos + constants.Shell.START_DISTANCE * self.tout,
                 )
 
-        if constants.Action.SNAP_BACK in self.actions:
-            self.snapping_back = True
+        temp_needs_update = self.__needs_update
+        self.__needs_update = False
+        return temp_needs_update
 
-        if constants.Action.STOP in self.actions and abs(self.speed) <= self.SNAP_STOP:
-            self.speed = 0.0
-
-        # snap_logic returns the number of degrees to turn this frame; all the other
-        # logic specifies the rate of turning in degrees per second
-        if not self.snapping_back:
-            ip_tangle *= delta
-        if not self.turning_back:
-            ip_bangle *= delta
-
-        # handle snapping/turning back
-        if self.snapping_back and ip_tangle == 0.0:
-            ip_tangle, self.snapping_back = self.snap_logic(
-                self.bangle, self.tangle, self.SNAP_SPEED * delta
-            )
-        if self.turning_back and ip_bangle == 0.0:
-            ip_bangle, self.turning_back = self.snap_logic(
-                self.tangle, self.bangle, self.BROTATE * delta
-            )
-
-        # adjust base and turret angles and out vectors if they've changed
-        if ip_bangle:
-            self.bangle += ip_bangle
-        if ip_tangle:
-            self.tangle += ip_tangle
-
-        # make sure the angles don't get too high, this helps the turret animation
-        self.bangle %= 360.0
-        self.tangle %= 360.0
-
-        # move the tank, according to the speed
-        self.pos += self.bout * self.speed * delta
-        # ensure the tank does not go over the edge of the world
-        self.pos[0] = max(min(self.pos[0], self.ground_hw), -self.ground_hw)
-        self.pos[2] = max(min(self.pos[2], self.ground_hw), -self.ground_hw)
-
-    @property
-    def state(self):
-        # TODO: make this dynamically change to not send already-known data
-        return {
-            "bangle": self.bangle,
-            "color": self.color,
-            "hits_left": self.hits_left,
-            "name": self.name,
-            "pos": tuple(self.pos),
-            "speed": self.speed,
-            "tangle": self.tangle,
-        }
-
-    # vector properties
-
-    # TODO: this is duplicated code from client.PlayerData - find a way to reuse it
-    @property
-    def bout(self):
-        # don't know if this is correct - will need some trial and error
-        return utils_3d.yaw(self.bangle, np.array((0.0, 0.0, 1.0)), np.array((1.0, 0.0, 0.0)))
-
-    @property
-    def tout(self):
-        # don't know if this is correct - will need some trial and error
-        return utils_3d.yaw(self.tangle, np.array((0.0, 0.0, 1.0)), np.array((1.0, 0.0, 0.0)))
-
-    # TODO: if these vectors are actually left instead of right, rename them
-    @property
-    def bright(self):
-        return utils_3d.normalize(np.cross(constants.UP, self.bout))
+    def update_actions(self, actions: set) -> None:
+        self.__needs_update = True
+        self.actions = actions
 
 
 class Server:
@@ -288,7 +132,7 @@ class Server:
         """Handle a message of type constants.Msg.REQUEST."""
         # Isn't it expensive to make new sets? Perhaps a new datatype should
         # be used for Tank.actions
-        self.tanks[client_id].actions = set(actions)
+        self.tanks[client_id].update_actions(set(actions))
 
     async def listen_for_start(self) -> None:
         """Start the game upon receiving proper user input."""
@@ -369,20 +213,21 @@ class Server:
         )
 
     async def send_updates(self) -> None:
-        # TODO: This will still cause an error for slow clients. This function should wait for all clients to send a message called STARTED.
-        await asyncio.sleep(0.05)
         while True:
             self.collisions()
             # TODO: maybe self.tanks should be list[tuple[int, Tank]] instead of dict[int, Tank]?
             for client_id, tank in self.tanks.items():
-                tank.update()
-                self.server.message_all(
-                    {
-                        "type": constants.Msg.APPROVE,
-                        "id": client_id,
-                        "state": tank.state,
-                    }
-                )
+                # tank.update() returns whether a network update is necessary
+                if tank.update():
+                    print(f"sending update for tank {id}")
+                    self.server.message_all(
+                        {
+                            "type": constants.Msg.APPROVE,
+                            "id": client_id,
+                            "state": tank.state,
+                        }
+                    )
+
             for shell in self.shells:
                 shell.update()
             for mine in self.mines:
